Revert glm::ortho with pixman
OpenGL applies the orthogonal projection as the last transformation to
map the points into [-1,1].

Unfortunately, we can't use this projection with pixman, but rather the
orthogonal projection that wlroots provides.

So here we keep all the OpenGL transformations, but then we revert the
orthogonal one to apply later the one from wlroots.