winshadow: improve shadow generation performance
Instead of computing the full shadows of the window, just compute the
top left shadow corner, and extend the values until getting half of the full
top left shadow. Then mirror it to get the other top right
shadow, and the top shadow is done.

Do the same for the left shadow.

The bottom shadow is also a mirror of the top shadow; same for the right
shadow.