winshadows: deinit_view when view is unmapped
So if the view we are attaching a winshadow is unmapped we
destroy all the shadow surface data. So if the surface is
reused because of a new map the surface data that could
have been already freed is not reused.

Synaptic package manager was displaying a Search window
dialog, than when closed was just hidden, so the widget
was still available but not visible. But when the search
window was re-opened it was just unhidden, but the
shadow surface data was not correctly initialized as it
was already been freed.